Refactor: Move WithTokenSpan and WithToken to own submodule
