+++ vhdl_lang/src/analysis/expression.rs
@@ -13,6 +13,7 @@ use super::overloaded::Disambiguated;
 use super::overloaded::DisambiguatedType;
 use super::overloaded::ResolvedCall;
 use super::scope::*;
+use crate::ast::token_range::{WithToken, WithTokenSpan};
 use crate::ast::*;
 use crate::data::error_codes::ErrorCode;
 use crate::data::*;
