+++ vhdl_lang/src/analysis/semantic.rs
@@ -8,6 +8,7 @@ use super::names::ResolvedName;
 use super::overloaded::Disambiguated;
 use super::overloaded::SubprogramKind;
 use super::scope::*;
+use crate::ast::token_range::WithTokenSpan;
 use crate::ast::*;
 use crate::data::error_codes::ErrorCode;
 use crate::data::*;
