+++ vhdl_lang/src/named_entity.rs
@@ -32,13 +32,13 @@ mod region;
 pub(crate) use region::RegionKind;
 pub use region::{AsUnique, NamedEntities, OverloadedName, Region, SetReference};
 mod formal_region;
+use crate::ast::token_range::WithToken;
 use crate::data::error_codes::ErrorCode;
 use crate::{TokenAccess, TokenSpan};
 pub use formal_region::{
     FormalRegion, GpkgInterfaceEnt, GpkgRegion, InterfaceClass, InterfaceEnt, RecordElement,
     RecordRegion,
 };
-use vhdl_lang::ast::WithToken;
 
 pub enum AnyEntKind<'a> {
     ExternalAlias {
