+++ vhdl_lang/src/syntax/expression.rs
@@ -8,8 +8,9 @@ use super::common::ParseResult;
 use super::names::{parse_name, parse_type_mark};
 use super::subtype_indication::parse_subtype_constraint;
 use super::tokens::{Kind, Kind::*};
+use crate::ast::token_range::{WithToken, WithTokenSpan};
 use crate::ast::{Literal, *};
-use crate::data::{Diagnostic, WithTokenSpan};
+use crate::data::Diagnostic;
 use crate::syntax::TokenAccess;
 use crate::{ast, TokenSpan};
 use vhdl_lang::syntax::parser::ParsingContext;
