+++ vhdl_lang/src/syntax/names.rs
@@ -11,9 +11,10 @@ use super::subprogram::parse_signature;
 use super::subtype_indication::parse_subtype_indication;
 use super::tokens::{Kind::*, TokenAccess};
 use crate::ast;
+use crate::ast::token_range::{WithToken, WithTokenSpan};
 use crate::ast::{Literal, *};
 use crate::data::error_codes::ErrorCode;
-use crate::data::{Diagnostic, WithTokenSpan};
+use crate::data::Diagnostic;
 use crate::syntax::separated_list::parse_list_with_separator_or_recover;
 use crate::syntax::TokenId;
 use vhdl_lang::syntax::parser::ParsingContext;
